refactor(reactor-api): rename adapter files to adapter-<type>-<concrete> convention
